trainds/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
PyJWT>=2.8.0
deep-translator>=1.11.0